Expose callback and run functions
